modal_app/requirements.txt
fastapi
numpy
pillow
tensorflow
